requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.0.8
django-crispy-forms==1.9.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
Werkzeug==1.0.1
